fixed migrations so they actually run now

+++ db/migrate/20200329021748_create_orders.rb
@@ -2,8 +2,8 @@ class CreateOrders < ActiveRecord::Migration[6.0]
   def change
     create_table :orders do |t|
       t.string :user_id
-      t.float :order_tax
-      t.int :order_total
+      t.decimal :order_tax
+      t.integer :order_total
 
       t.timestamps
     end
