+++ db/migrate/20200329022006_create_provinces.rb
@@ -2,7 +2,7 @@ class CreateProvinces < ActiveRecord::Migration[6.0]
   def change
     create_table :provinces do |t|
       t.string :name
-      t.float :tax
+      t.decimal :tax
 
       t.timestamps
     end
